requirements.txt
asyncpg
python-dotenv
aiofiles
aiohttp
orjson
discord.py